# index.js
const bodyParser = require("body-parser");
const config = require("./config.json");
const express = require("express");
const cors = require("cors");
const logger = require("morgan");
const chalk = require("chalk");
const Bard = require("./bard");
const axios = require("axios");
const app = express();
const PORT = 7860;
const swaggerDocument = require("./swagger.json");
const swaggerUi = require("swagger-ui-express");

app.set("json spaces", 2);
app.set("trust proxy", true);
app.enable("trust proxy");
app.use(bodyParser.json({ limit: "50mb" }));
app.use(bodyParser.urlencoded({ limit: "50mb", extended: false }));
app.use(swaggerUi.serve);
app.use(cors());

function formatBytes(bytes, decimals = 2) {
  if (bytes === 0) return "0 Bytes";
  const k = 1024;
  const dm = decimals < 0 ? 0 : decimals;
  const sizes = ["Bytes", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"];
  const i = Math.floor(Math.log(bytes) / Math.log(k));
  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + " " + sizes[i];
}

function status(code) {
  if (code > 400 && code < 499) return chalk.yellow(code);
  if (code > 500 && code < 599) return chalk.red(code);
  if (code > 299 && code < 399) return chalk.cyan(code);
  if (code > 199) return chalk.green(code);
  return chalk.yellow(code);
}

app.use(
  logger(function (tokens, req, res) {
    return (
      "[ ✨ ] " +
      [
        req.ip,
        tokens.method(req, res),
        tokens.url(req, res),
        status(tokens.status(req, res)),
        tokens["response-time"](req, res) + " ms",
        formatBytes(
          isNaN(tokens.res(req, res, "content-length"))
            ? 0
            : tokens.res(req, res, "content-length"),
        ),
      ].join(" | ")
    );
  }),
);

app.all(["/backend/conversation"], (req, res, next) => {
  if (req.method !== "POST") {
    return res.status(405).json({
      content: "Method not allowed",
      status: 405,
      creator: "RizzyFuzz",
    });
  }
  next();
});

app.post("/backend/conversation", async (req, res) => {
  try {
    const { ask, image, preview } = req.body;
    if (!ask) {
      return res.status(400).json({
        content: "Bad Request: No Query Ask Provided",
        status: 400,
        creator: "RizzyFuzz",
      });
    }

    const bard = new Bard();
    await bard.configure(config.cookie);

    let { status, content } = await bard.initial(ask, image || null);

    if (!status) {
      return res.status(500).json({
        content,
        status: 500,
        creator: "RizzyFuzz",
      });
    }

    if (preview && typeof content === "string" && content.startsWith("https")) {
      try {
        const hdUrl = content.replace(/=s\d+$/, "=s1024");
        const response = await axios.get(hdUrl, {
          responseType: "arraybuffer",
          headers: {
            accept: "image/avif,image/webp,image/apng,image/*,*/*;q=0.8",
            referer: "https://gemini.google.com/",
            "user-agent":
              "Mozilla/5.0 (Linux; Android 6.0; Nexus 5 Build/MRA58N) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/140.0.0.0 Mobile Safari/537.36",
          },
        });

        res.setHeader(
          "Content-Type",
          response.headers["content-type"] || "image/png",
        );
        return res.send(Buffer.from(response.data, "binary"));
      } catch (err) {
        return res.status(500).json({
          content: "Failed to fetch image for preview",
          status: 500,
          creator: "RizzyFuzz",
        });
      }
    }
    content = content.replace(/=s\d+$/, "=s1024");
    res.status(200).json({ content, status: 200, creator: "RizzyFuzz" });
  } catch (error) {
    console.error(error);
    res.status(500).json({
      content: "Internal Server Error!",
      status: 500,
      creator: "RizzyFuzz",
    });
  }
});

app.get("/", (req, res) => {
  swaggerDocument.host = req.get("host");
  swaggerDocument.schemes = [config.schemes];
  res.send(
    swaggerUi.generateHTML(swaggerDocument, {
      customCss: `.swagger-ui .topbar .download-url-wrapper { display: none } 
    .swagger-ui .topbar-wrapper img[alt="Bardie API"], .topbar-wrapper span {
      visibility: collapse;
    }
    .swagger-ui .topbar-wrapper img {
      content: url("https://www.gstatic.com/lamda/images/favicon_v1_150160cddff7f294ce30.svg");
    }
    .swagger-ui .opblock-section-body .parameters-col_description {
     width: 50px;
    }`,
      customfavIcon:
        "https://www.gstatic.com/lamda/images/favicon_v1_150160cddff7f294ce30.svg",
      customSiteTitle: swaggerDocument.info.title,
      customSiteDesc: swaggerDocument.info.description,
    }),
  );
});

app.get("/swagger.json", (req, res) => {
  swaggerDocument.host = req.get("host");
  swaggerDocument.schemes = [config.schemes];
  res.json(swaggerDocument);
});

app.use((req, res, next) => {
  res.status(404).json({
    content: "Not Found!",
    status: 404,
    creator: "RizzyFuzz",
  });
});

app.listen(PORT, () => {
  console.log("Server running on port " + PORT);
});

module.exports = app;